[ADD] fscore, precision, recall, accuracy, wall clock time, FLOPs, saved models and outputs of 3 CNNs
