+++ requirements.txt
@@ -16,3 +16,6 @@ matplotlib==3.8.0
 
 # Data loading and manipulation
 pandas==2.1.1
+
+# FLOPs calculation
+fvcore
