+++ src/evaluation/evaluate.py
@@ -5,7 +5,7 @@ from torch.utils.data import DataLoader, TensorDataset
 import sys
 import os
 sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "../..")))
-from sklearn.metrics import confusion_matrix, ConfusionMatrixDisplay
+from sklearn.metrics import confusion_matrix, ConfusionMatrixDisplay, classification_report, accuracy_score
 from src.models.cnn_models import build_resnet_model
 from src.preprocessing.constants import *
 
@@ -111,6 +111,13 @@ def plot_confusion_matrix(y_true, y_pred, class_names):
     disp.plot(cmap="Blues", xticks_rotation=45)
     plt.title("Confusion Matrix")
     plt.show()
+
+def evaluate_scores(y_true, y_pred, class_names):
+    report = classification_report(y_true, y_pred, target_names=DATASET_CLASSES[dataset_type], digits=4)
+    accuracy = accuracy_score(y_true, y_pred)
+    print("\nClassification Report:")
+    print(report)
+    print(f"Overall Test Accuracy: {accuracy * 100:.2f}%")
