+++ src/models/train_DenseNet.py
@@ -5,15 +5,19 @@ import torch.nn as nn
 import torch.optim as optim
 from torchvision import models
 import numpy as np
-sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "../..")))
-from train_ResNet import save_model,evaluate_on_test_data, load_data
-from src.preprocessing.constants import *
-from src.evaluation.evaluate import load_model,plot_metrics,get_predictions,plot_confusion_matrix
+import time
 
-dataset_type = DatasetType.LARGE #.LARGE, .MODERATE, .SMALL
+sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "../..")))
+from train_ResNet import save_model, evaluate_on_test_data, load_data
+from src.preprocessing.constants import *
+from src.evaluation.evaluate import load_model, plot_metrics, get_predictions, plot_confusion_matrix, evaluate_scores
+from fvcore.nn import FlopCountAnalysis
+
+
+dataset_type = DatasetType.SMALL  # .LARGE, .MODERATE, .SMALL
 
 class DenseNetModel:
-    def __init__(self, num_classes, pretrained=True):
+    def __init__(self, num_classes, pretrained=False):
         self.model = self._get_model(num_classes, pretrained)
 
     def _get_model(self, num_classes, pretrained):
@@ -21,7 +25,7 @@ class DenseNetModel:
             model = models.densenet121(weights="IMAGENET1K_V1")
         else:
             model = models.densenet121(weights=None)
-        
+
         # Modify the classifier to match the number of classes
         model.classifier = nn.Linear(model.classifier.in_features, num_classes)
         return model
@@ -30,20 +34,22 @@ class DenseNetModel:
         return self.model
 
 
-def train(model, train_loader, val_loader, criterion, optimizer, device, num_epochs=10):
+def train(model, train_loader, val_loader, criterion, optimizer, device, num_epochs=12):
     model.to(device)
 
     train_losses, val_losses = [], []
     train_accuracies, val_accuracies = [], []
 
     for epoch in range(num_epochs):
+        start_time = time.time()
+
         model.train()
         running_loss = 0.0
         correct = 0
         total = 0
 
-        print(f"\nStarting Epoch {epoch + 1}/{num_epochs}")
-        print('Training...')
+        # print(f"\nStarting Epoch {epoch + 1}/{num_epochs}")
+        # print('Training...')
 
         for batch_idx, (inputs, labels) in enumerate(train_loader):
             inputs, labels = inputs.to(device), labels.to(device)
@@ -63,14 +69,17 @@ def train(model, train_loader, val_loader, criterion, optimizer, device, num_epo
             correct += (predicted == labels).sum().item()
 
             # Log progress within each epoch (every 10 batches)
-            if (batch_idx + 1) % 10 == 0:
-                print(f"  Batch {batch_idx + 1}/{len(train_loader)}, "
-                      f"Train Loss: {loss.item():.4f}, "
-                      f"Train Accuracy (Running): {correct / total:.4f}")
+            # if (batch_idx + 1) % 10 == 0:
+            #     print(f"  Batch {batch_idx + 1}/{len(train_loader)}, "
+            #           f"Train Loss: {loss.item():.4f}, "
+            #           f"Train Accuracy (Running): {correct / total:.4f}")
 
         # Calculate epoch's train loss and accuracy
         epoch_train_loss = running_loss / total
         epoch_train_accuracy = correct / total
+
+        end_time = time.time()
+        epoch_time = end_time - start_time
 
         train_losses.append(epoch_train_loss)
         train_accuracies.append(epoch_train_accuracy)
@@ -80,11 +89,14 @@ def train(model, train_loader, val_loader, criterion, optimizer, device, num_epo
         val_losses.append(val_loss)
         val_accuracies.append(val_accuracy)
 
+
+
         print(f"Epoch {epoch + 1}/{num_epochs} Completed - "
               f"Train Loss: {epoch_train_loss:.4f}, "
               f"Train Accuracy: {epoch_train_accuracy:.4f}, "
               f"Validation Loss: {val_loss:.4f}, "
-              f"Validation Accuracy: {val_accuracy:.4f}")
+              f"Validation Accuracy: {val_accuracy:.4f}, "
+              f"Wall Clock Time: {epoch_time:.2f} s")
 
     return train_losses, val_losses, train_accuracies, val_accuracies
 
@@ -96,7 +108,7 @@ def evaluate(model, val_loader, criterion):
     total = 0
 
     with torch.no_grad():
-        print('Validating...')
+        # print('Validating...')
         for i, (inputs, labels) in enumerate(val_loader):
             inputs, labels = inputs.to(device), labels.to(device)
             outputs = model(inputs)
@@ -107,15 +119,19 @@ def evaluate(model, val_loader, criterion):
             total += labels.size(0)
             correct += (predicted == labels).sum().item()
 
-            if (i + 1) % 10 == 0:  # Log every 10 batches
-                print(f"Batch {i + 1}/{len(val_loader)}, "
-                      f"Validation Loss: {loss.item():.4f}, "
-                      f"Validation Accuracy: {correct / total:.4f}")
+            # if (i + 1) % 10 == 0:  # Log every 10 batches
+            #     print(f"Batch {i + 1}/{len(val_loader)}, "
+            #           f"Validation Loss: {loss.item():.4f}, "
+            #           f"Validation Accuracy: {correct / total:.4f}")
 
     val_loss /= total
     val_accuracy = correct / total
-    print(f"Validation Loss: {val_loss:.4f}, Validation Accuracy: {val_accuracy:.4f}")
     return val_loss, val_accuracy
+
+def flop_calculator(model):
+    input_tensor = torch.randn(32, 3, 224, 224).to(device)
+    flops = FlopCountAnalysis(model, input_tensor)
+    print(f"Total FLOPs: {flops.total()} FLOPs")
 
 
 if __name__ == "__main__":
@@ -124,27 +140,28 @@ if __name__ == "__main__":
     num_classes = len(DATASET_CLASSES[dataset_type])
     num_epochs = 10
     learning_rate = 0.001
-    path = os.path.join(os.getcwd(), 'trained models/pretrained weights/malware_'+dataset_type.name.lower()+'_densenet121.pth')
+    path = os.path.join(os.getcwd(),
+                        'trained models/no pretrained weights/malware_' + dataset_type.name.lower() + '_densenet121.pth')
 
     # Load data
-    train_loader, val_loader,test_loader = load_data(foldername)
+    train_loader, val_loader, test_loader = load_data(foldername)
 
     # Initialize the DenseNet-121 model
     densenet_model = DenseNetModel(num_classes=num_classes)
     model = densenet_model.get_model().to(device)
-
 
     # Define loss function and optimizer
     criterion = nn.CrossEntropyLoss()
     optimizer = optim.Adam(model.parameters(), lr=learning_rate)
 
     # Training loop
-    train_losses, val_losses, train_accuracies, val_accuracies = train(model, train_loader, val_loader, criterion,optimizer,device,num_epochs)
+    train_losses, val_losses, train_accuracies, val_accuracies = train(model, train_loader, val_loader, criterion,
+                                                                       optimizer, device, num_epochs)
 
     save_model(model, path)
 
     trained_model = load_model(path, len(DATASET_CLASSES[dataset_type]))
-    
+
     plot_metrics(train_losses, val_losses, train_accuracies, val_accuracies)
 
     test_loss, test_accuracy = evaluate_on_test_data(model, test_loader, criterion)
@@ -158,6 +175,11 @@ if __name__ == "__main__":
     print("Unique classes in y_true:", unique_true)
     print("Unique classes in y_pred:", unique_pred)
 
+    # FLOPs
+    flop_calculator(model)
+
     # Plot confusion matrix
     plot_confusion_matrix(y_true, y_pred, DATASET_CLASSES[dataset_type])
 
+    # Precision, F1-score, Recall & Accuracy
+    evaluate_scores(y_true, y_pred, DATASET_CLASSES[dataset_type])
