+++ src/models/train_EfficientNet.py
@@ -5,16 +5,19 @@ import numpy as np
 import torch.nn as nn
 import torch.optim as optim
 from torchvision import models
+import time
+
+from fvcore.nn import FlopCountAnalysis
 sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "../..")))
 from train_ResNet import save_model, evaluate_on_test_data, load_data, device
 from src.preprocessing.constants import *
-from src.evaluation.evaluate import load_model,plot_metrics,get_predictions,plot_confusion_matrix
+from src.evaluation.evaluate import load_model, plot_metrics, get_predictions, plot_confusion_matrix, evaluate_scores
 
+dataset_type = DatasetType.SMALL
 
-dataset_type = DatasetType.LARGE
 
 class EfficientNetModel:
-    def __init__(self, num_classes, pretrained=True):
+    def __init__(self, num_classes, pretrained=False):
         self.model = self._get_model(num_classes, pretrained)
 
     def _get_model(self, num_classes, pretrained):
@@ -29,15 +32,18 @@ class EfficientNetModel:
 
     def get_model(self):
         return self.model
-    
+
+
 # Define the EfficientNet-B0 model
-def train(model, train_loader,val_loader,criterion, optimizer, num_epochs):
+def train(model, train_loader, val_loader, criterion, optimizer, num_epochs):
     model.to(device)
 
     train_losses, val_losses = [], []
     train_accuracies, val_accuracies = [], []
 
     for epoch in range(num_epochs):
+        start_time = time.time()
+
         model.train()
         running_loss = 0.0
         correct = 0
@@ -61,15 +67,21 @@ def train(model, train_loader,val_loader,criterion, optimizer, num_epochs):
         epoch_loss = running_loss / total
         epoch_accuracy = correct / total
 
+        end_time = time.time()
+        epoch_time = end_time - start_time
+
         train_losses.append(epoch_loss)
         train_accuracies.append(epoch_accuracy)
         val_loss, val_accuracy = evaluate(model, val_loader, criterion)
         val_losses.append(val_loss)
         val_accuracies.append(val_accuracy)
 
-        print(f"Epoch {epoch+1}/{num_epochs}, Train Loss: {epoch_loss:.4f}, Train Accuracy: {epoch_accuracy:.4f}, Validation Loss: {val_loss:.4f}, Validation Accuracy: {val_accuracy:.4f}")
+
+        print(
+            f"Epoch {epoch + 1}/{num_epochs}, Train Loss: {epoch_loss:.4f}, Train Accuracy: {epoch_accuracy:.4f}, Validation Loss: {val_loss:.4f}, Validation Accuracy: {val_accuracy:.4f}, Wall Clock Time: {epoch_time:.2f} s")
 
     return train_losses, val_losses, train_accuracies, val_accuracies
+
 
 # Evaluation function
 def evaluate(model, val_loader, criterion):
@@ -93,28 +105,36 @@ def evaluate(model, val_loader, criterion):
     val_accuracy = correct / total
     return val_loss, val_accuracy
 
+def flop_calculator(model):
+    input_tensor = torch.randn(32, 3, 224, 224).to(device)
+    flops = FlopCountAnalysis(model, input_tensor)
+    print(f"Total FLOPs: {flops.total()} FLOPs")
+
+
 if __name__ == "__main__":
     foldername = 'malware-images-' + dataset_type.name.lower()
 
     num_classes = len(DATASET_CLASSES[dataset_type])
     batch_size = 32
-    num_epochs = 10
+    num_epochs = 12
     learning_rate = 0.001
 
     # Load data
-    train_loader, val_loader,test_loader = load_data(foldername)
+    train_loader, val_loader, test_loader = load_data(foldername)
 
     # Initialize the model
     efficientnet_model = EfficientNetModel(num_classes=num_classes)
     model = efficientnet_model.get_model().to(device)
 
-    path = os.path.join(os.getcwd(), 'trained models/pretrained weights/malware_'+dataset_type.name.lower()+'_efficientnetb0.pth')
+    path = os.path.join(os.getcwd(),
+                        'trained models/no pretrained weights/malware_' + dataset_type.name.lower() + '_efficientnetb0.pth')
 
     # Define loss function and optimizer
     criterion = nn.CrossEntropyLoss()
     optimizer = optim.Adam(model.parameters(), lr=learning_rate)
 
-    train_losses, val_losses, train_accuracies, val_accuracies = train(model, train_loader, val_loader, criterion, optimizer, num_epochs)
+    train_losses, val_losses, train_accuracies, val_accuracies = train(model, train_loader, val_loader, criterion,
+                                                                       optimizer, num_epochs)
 
     save_model(model, path)
 
@@ -135,8 +155,11 @@ if __name__ == "__main__":
     print("Unique classes in y_true:", unique_true)
     print("Unique classes in y_pred:", unique_pred)
 
+    # FLOPs
+    flop_calculator(model)
+
     # Plot confusion matrix
     plot_confusion_matrix(y_true, y_pred, DATASET_CLASSES[dataset_type])
 
-
-
+    # Precision, F1-score, Recall & Accuracy
+    evaluate_scores(y_true, y_pred, DATASET_CLASSES[dataset_type])
