+++ src/models/train_ResNet.py
@@ -3,23 +3,26 @@ import os
 import torch
 import torch.nn as nn
 import torch.optim as optim
-from torch.utils.data import DataLoader, TensorDataset
 import numpy as np
+import time
+
+from fvcore.nn import FlopCountAnalysis
+from torch.utils.data import DataLoader, TensorDataset
 from torchvision.models import resnet18, ResNet18_Weights
 sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "../..")))
 from src.preprocessing.constants import *
-from src.evaluation.evaluate import load_model,plot_metrics,get_predictions,plot_confusion_matrix
+from src.evaluation.evaluate import load_model, plot_metrics, get_predictions, plot_confusion_matrix, evaluate_scores
 
 
 device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
 
-dataset_type = DatasetType.LARGE #.LARGE, .MODERATE, .SMALL
+dataset_type = DatasetType.SMALL #.LARGE, .MODERATE, .SMALL
 
 class ResNet18(nn.Module):
     def __init__(self, num_classes):
         super(ResNet18, self).__init__()
-        self.model = resnet18(weights=ResNet18_Weights.IMAGENET1K_V1)
-        # self.model = resnet18(weights=None)
+        # self.model = resnet18(weights=ResNet18_Weights.IMAGENET1K_V1)
+        self.model = resnet18(weights=None)
         self.model.fc =  nn.Sequential(nn.Linear(self.model.fc.in_features, num_classes))
 
     def forward(self, x):
@@ -131,7 +134,7 @@ def validate_model(model, val_loader, criterion):
     return val_loss, val_accuracy
 
 
-def train_loop(model,train_loader,val_loader,criterion,num_epochs=10):
+def train_loop(model,train_loader,val_loader,criterion,num_epochs=8):
 
     optimizer = optim.Adam(model.parameters(), lr=1e-4)
 
@@ -141,8 +144,13 @@ def train_loop(model,train_loader,val_loader,criterion,num_epochs=10):
     y_val_true, y_val_pred = [], []
 
     for epoch in range(num_epochs):
+        start_time = time.time()
+
         train_loss, train_accuracy = train_model(model, train_loader, optimizer, criterion)
         val_loss, val_accuracy = validate_model(model, val_loader, criterion)
+
+        end_time = time.time()
+        epoch_time = end_time - start_time
 
         train_losses.append(train_loss)
         val_losses.append(val_loss)
@@ -150,8 +158,9 @@ def train_loop(model,train_loader,val_loader,criterion,num_epochs=10):
         val_accuracies.append(val_accuracy)
 
         print(f"Epoch {epoch + 1}/{num_epochs}, "
-              f"Train Loss: {train_loss:.4f}, Train Accuracy: {train_accuracy:.4f}, "
-              f"Val Loss: {val_loss:.4f}, Val Accuracy: {val_accuracy:.4f}")
+              f"Train Loss: {train_loss:.4f}, Train Accuracy: {train_accuracy * 100:.2f}%, "
+              f"Val Loss: {val_loss:.4f}, Val Accuracy: {val_accuracy * 100:.2f}%, "
+              f"Wall Clock Time: {epoch_time:.2f} s")
 
     return train_losses, val_losses, train_accuracies, val_accuracies
 
@@ -193,6 +202,11 @@ def save_model(model, path="model.pth"):
     torch.save(model.state_dict(), path)
     print(f"Model saved to {path}")
 
+def flop_calculator(model):
+    input_tensor = torch.randn(32, 3, 224, 224).to(device)
+    flops = FlopCountAnalysis(model, input_tensor)
+    print(f"Total FLOPs: {flops.total()} FLOPs")
+
 
 # Run the main function
 if __name__ == "__main__":
@@ -207,7 +221,7 @@ if __name__ == "__main__":
 
     train_losses, val_losses, train_accuracies, val_accuracies=train_loop(model,train_loader,val_loader,criterion)
 
-    path = os.path.join(os.getcwd(),'trained models/pretrained weights/malware_'+dataset_type.name.lower()+'_resnet18.pth')
+    path = os.path.join(os.getcwd(),'trained models/no pretrained weights/malware_'+dataset_type.name.lower()+'_resnet18.pth')
 
 
     save_model(model, path)
@@ -229,5 +243,11 @@ if __name__ == "__main__":
     print("Unique classes in y_true:", unique_true)
     print("Unique classes in y_pred:", unique_pred)
 
+    # FLOPs
+    flop_calculator(model)
+
     # Plot confusion matrix
     plot_confusion_matrix(y_true, y_pred, DATASET_CLASSES[dataset_type])
+
+    #Precision, F1-score, Recall & Accuracy
+    evaluate_scores(y_true, y_pred, DATASET_CLASSES[dataset_type])
