+++ src/preprocessing/data_preprocessing_Transfer_learning.py
@@ -1,0 +1,145 @@
+import sys
+import numpy as np
+from utils import *
+sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "../..")))
+from src.preprocessing.constants import *
+
+import torch
+from torchvision import transforms
+from PIL import Image
+
+augmentation_transforms = transforms.Compose([
+    transforms.RandomHorizontalFlip(),
+    transforms.RandomVerticalFlip(),
+    transforms.RandomRotation(30),
+    transforms.ColorJitter(brightness=0.2, contrast=0.2, saturation=0.2, hue=0.2)
+])
+
+def augment_data(data, transformations, target_size=(128, 128)):
+    augmented_data = {}
+    for class_name, images in data.items():
+        augmented_images = []
+        for image in images:
+            pil_image = Image.fromarray((image * 255).astype(np.uint8))
+            augmented_image = transformations(pil_image)
+            augmented_image_np = np.array(augmented_image) / 255.0
+            augmented_images.append(augmented_image_np.astype(np.float32))
+        augmented_data[class_name] = augmented_images
+    return augmented_data
+
+# # Loop through each DatasetType (SMALL, MODERATE, LARGE)
+# for dataset_type in DatasetType:
+#     print(f"\nProcessing dataset type: {dataset_type.name}")
+#
+#     # Load images based on the current dataset type
+#     data = load_images(DATA_DIR + dataset_type.value, DATASET_CLASSES[dataset_type])
+#
+#     # Check if total image count exceeds the maximum allowed, and limit if necessary
+#     total_image_count = sum(len(images) for images in data.values())
+#     if total_image_count > MAX_TOTAL_IMAGES:
+#         print(f"Total images ({total_image_count}) exceed {MAX_TOTAL_IMAGES}. Limiting count per class.")
+#         # Calculate target count per class to maintain equal distribution under the limit
+#         target_count_per_class = MAX_TOTAL_IMAGES // len(DATASET_CLASSES[dataset_type])
+#         data = limit_image_count(data, target_count_per_class)
+
+
+
+MAX_TOTAL_IMAGES=50000
+#set dataset type here for performing preprocessing on
+dataset_type = DatasetType.LARGE #.LARGE, .MODERATE, .SMALL
+
+# Load images
+data = load_images('C:/data/raw/malware_dataset/train', DATASET_CLASSES[dataset_type])
+
+# Limit image count dynamically to ensure total stays under 50K
+data = limit_image_count(data, max_total_images=40000)
+
+
+# Check if total image count exceeds the maximum allowed, and limit if necessary
+total_image_count = sum(len(images) for images in data.values())
+if total_image_count > MAX_TOTAL_IMAGES:
+    print(f"Total images ({total_image_count}) exceed {MAX_TOTAL_IMAGES}. Limiting count per class.")
+    # Calculate target count per class to maintain equal distribution under the limit
+    target_count_per_class = MAX_TOTAL_IMAGES // len(DATASET_CLASSES[dataset_type])
+    data = limit_image_count(data, target_count_per_class)
+
+
+# Check the number of images per class
+for class_name, images in data.items():
+    print(f"Class '{class_name}' has {len(images)} images.")
+
+# Display a few images from each class for inspection
+#test_display(DATASET_CLASSES[dataset_type],data)
+
+# Increase dataset, three times
+augmented_data = augment_data(data, augmentation_transforms, target_size=(128, 128))
+augmented_data2 = augment_data(data, augmentation_transforms, target_size=(128, 128))
+augmented_data3 = augment_data(data, augmentation_transforms, target_size=(128, 128))
+augmented_data4 = augment_data(data, augmentation_transforms, target_size=(128, 128))
+data = {class_name: data[class_name] + augmented_data[class_name] for class_name in data.keys()}
+data = {class_name: data[class_name] + augmented_data2[class_name] for class_name in data.keys()}
+data = {class_name: data[class_name] + augmented_data3[class_name] for class_name in data.keys()}
+data = {class_name: data[class_name] + augmented_data4[class_name] for class_name in data.keys()}
+
+
+target_sample_count = calculate_target_sample_count(data)
+
+# Apply balancing with updated target sample count
+balanced_images_array, balanced_labels_array = balance_classes(data, target_sample_count)
+print("balanced images", balanced_images_array.shape)
+
+#Normalize Data
+normalized_images_array = normalize_images(balanced_images_array)
+print("normalized images:")
+print(normalized_images_array[:1])
+
+encoded_labels, label_mapping = encode_labels(balanced_labels_array)
+print("encoded labels", encoded_labels)
+
+# If you want a test split as well, you can set test_size
+X_train, y_train = make_training_data(normalized_images_array, encoded_labels, 'malware-images-'+dataset_type.name.lower())
+
+print("Class distribution in Training Set:")
+unique_train, counts_train = np.unique(y_train, return_counts=True)
+for class_label, count in zip(unique_train, counts_train):
+    print(f"Class {class_label}: {count} samples")
+
+# Load validation images to use as test set
+data = load_images('C:/data/raw/malware_dataset/val', DATASET_CLASSES[dataset_type])
+# Limit image count dynamically to ensure total stays under 50K
+data = limit_image_count(data, max_total_images=40000)
+# Check if total image count exceeds the maximum allowed, and limit if necessary
+total_image_count = sum(len(images) for images in data.values())
+if total_image_count > MAX_TOTAL_IMAGES:
+    print(f"Total images ({total_image_count}) exceed {MAX_TOTAL_IMAGES}. Limiting count per class.")
+    # Calculate target count per class to maintain equal distribution under the limit
+    target_count_per_class = MAX_TOTAL_IMAGES // len(DATASET_CLASSES[dataset_type])
+    data = limit_image_count(data, target_count_per_class)
+
+# Check the number of images per class
+for class_name, images in data.items():
+    print(f"Class '{class_name}' has {len(images)} images.")
+# Display a few images from each class for inspection
+#test_display(DATASET_CLASSES[dataset_type], data)
+
+target_sample_count = calculate_target_sample_count(data)
+# Apply balancing with updated target sample count
+balanced_images_array, balanced_labels_array = balance_classes(data, target_sample_count)
+print("balanced images", balanced_images_array.shape)
+# Normalize Data
+normalized_images_array = normalize_images(balanced_images_array)
+print("normalized images", normalized_images_array)
+encoded_labels, label_mapping = encode_labels(balanced_labels_array)
+print("encoded labels", encoded_labels)
+# Save the validation images as test set
+X_test, X_val, y_test, y_val = test_val_split(normalized_images_array, encoded_labels, 'malware-images-'+dataset_type.name.lower(), val_size=0.2, random_state=0 )
+
+print("\nClass distribution in Validation Set:")
+unique_val, counts_val = np.unique(y_val, return_counts=True)
+for class_label, count in zip(unique_val, counts_val):
+    print(f"Class {class_label}: {count} samples")
+
+print("\nClass distribution in Test Set:")
+unique_test, counts_test = np.unique(y_test, return_counts=True)
+for class_label, count in zip(unique_test, counts_test):
+    print(f"Class {class_label}: {count} samples")
