--- tran_EfficientNet_moderate_to_large.py
@@ -1,217 +1,0 @@
-import os
-import sys
-import torch
-import numpy as np
-import torch.nn as nn
-import torch.optim as optim
-from torch.utils.data import DataLoader, TensorDataset
-from torchvision import models
-
-sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "../..")))
-from train_ResNet import save_model, evaluate_on_test_data, device
-from src.preprocessing.constants import *
-from src.evaluation.evaluate import load_model, plot_metrics, get_predictions, plot_confusion_matrix
-
-dataset_type = DatasetType.LARGE
-
-
-class EfficientNetModel:
-    def __init__(self, num_classes, pretrained=True, weight_path=None):
-        self.model = self._get_model(num_classes, pretrained, weight_path)
-
-    def _get_model(self, num_classes, pretrained, weight_path):
-        # Initialize the model (without pre-trained models weights)
-        model = models.efficientnet_b0(weights=None)
-        # Modify the classifier to match the number of classes
-        model.classifier[1] = nn.Linear(model.classifier[1].in_features, num_classes)
-        # If weights are provided, load them
-        if weight_path and os.path.exists(weight_path):
-            print(f"Loading weights from {weight_path}")
-            state_dict = torch.load(weight_path, map_location=torch.device('cpu'))
-            # Remove incompatible layers
-            state_dict = {k: v for k, v in state_dict.items() if "classifier.1" not in k}
-
-            # Load the state_dict into the model
-            missing, unexpected = model.load_state_dict(state_dict, strict=False)
-
-            # Log any missing or unexpected keys
-            print("Loaded layers:", state_dict.keys())
-            print("Missing layers:", missing)
-            print("Unexpected layers:", unexpected)
-            print("Weights loaded successfully, excluding incompatible layers.")
-
-
-        return model
-
-    def get_model(self):
-        return self.model
-
-
-def load_data(foldername):
-    ABSOLUTE_DATA_PATH = "C:/Users/zheng/OneDrive/Desktop/CurrentSem/COMP6721(AppliedAI/Project/Malware-Detection-CNN/src/preprocessing/data/"
-    # Construct absolute paths for the dataset files
-    data_path = os.path.join(ABSOLUTE_DATA_PATH, 'processed', foldername)
-    print("data_path: "+data_path)
-    # Load the data using the absolute path
-    X_train = np.load('C:/Users/zheng/OneDrive/Desktop/CurrentSem/COMP6721(AppliedAI/Project/Malware-Detection-CNN/src/preprocessing/data/processed/malware-images-large/X_train.npy')
-    y_train = np.load('C:/Users/zheng/OneDrive/Desktop/CurrentSem/COMP6721(AppliedAI/Project/Malware-Detection-CNN/src/preprocessing/data/processed/malware-images-large/y_train.npy')
-    X_val = np.load('C:/Users/zheng/OneDrive/Desktop/CurrentSem/COMP6721(AppliedAI/Project/Malware-Detection-CNN/src/preprocessing/data/processed/malware-images-large/X_val.npy')
-    y_val = np.load('C:/Users/zheng/OneDrive/Desktop/CurrentSem/COMP6721(AppliedAI/Project/Malware-Detection-CNN/src/preprocessing/data/processed/malware-images-large/y_val.npy')
-
-    # Stack the grayscale image data into 3 channels to match ResNet-18’s expected input format
-    X_train_tensor = torch.tensor(X_train, dtype=torch.float32).unsqueeze(1).repeat(1, 3, 1,1)  # (num_images, 3, 224, 224)
-    y_train_tensor = torch.tensor(y_train, dtype=torch.long)
-
-    X_val_tensor = torch.tensor(X_val, dtype=torch.float32).unsqueeze(1).repeat(1, 3, 1,1)  # (num_images_val, 3, 224, 224)
-    y_val_tensor = torch.tensor(y_val, dtype=torch.long)
-
-    train_dataset = TensorDataset(X_train_tensor, y_train_tensor)
-    val_dataset = TensorDataset(X_val_tensor, y_val_tensor)
-
-    train_loader = DataLoader(train_dataset, batch_size=32, shuffle=True)
-    val_loader = DataLoader(val_dataset, batch_size=32, shuffle=False)
-
-    X_test = np.load('C:/Users/zheng/OneDrive/Desktop/CurrentSem/COMP6721(AppliedAI/Project/Malware-Detection-CNN/src/preprocessing/data/processed/malware-images-large/X_test.npy')
-    y_test = np.load('C:/Users/zheng/OneDrive/Desktop/CurrentSem/COMP6721(AppliedAI/Project/Malware-Detection-CNN/src/preprocessing/data/processed/malware-images-large/y_test.npy') # Adjust path to your .npy file
-
-    # Convert numpy arrays to PyTorch tensors
-    X_test_tensor = torch.tensor(X_test, dtype=torch.float32).unsqueeze(1).repeat(1, 3, 1, 1)
-    y_test_tensor = torch.tensor(y_test, dtype=torch.long)
-
-    # Move the test data to the same device
-    X_test_tensor = X_test_tensor.to(device)
-    y_test_tensor = y_test_tensor.to(device)
-
-    test_data = TensorDataset(X_test_tensor, y_test_tensor)
-    test_loader = DataLoader(test_data, batch_size=32, shuffle=False)
-
-
-    # Check the count of each class based on y_test labels
-    unique_classes, class_counts = np.unique(y_test, return_counts=True)
-
-    print("Class distribution in X_test based on y_test labels:")
-    for cls, count in zip(unique_classes, class_counts):
-        print(f"Class {cls}: {count} samples in X_test")
-
-    unique_classes_train, class_counts_train = np.unique(y_train, return_counts=True)
-
-    print("Class distribution in X_train based on y_train labels:")
-    for cls, count in zip(unique_classes_train, class_counts_train):
-        print(f"Class {cls}: {count} samples in X_train")
-    return train_loader, val_loader,test_loader
-
-# Define the EfficientNet-B0 model
-def train(model, train_loader, val_loader, criterion, optimizer, num_epochs):
-    model.to(device)
-
-    train_losses, val_losses = [], []
-    train_accuracies, val_accuracies = [], []
-
-    for epoch in range(num_epochs):
-        model.train()
-        running_loss = 0.0
-        correct = 0
-        total = 0
-        print(f'Epoch: {epoch} started')
-        for inputs, labels in train_loader:
-            inputs, labels = inputs.to(device), labels.to(device)
-
-            optimizer.zero_grad()
-            outputs = model(inputs)
-            loss = criterion(outputs, labels)
-
-            loss.backward()
-            optimizer.step()
-
-            running_loss += loss.item() * inputs.size(0)
-            _, predicted = torch.max(outputs, 1)
-            total += labels.size(0)
-            correct += (predicted == labels).sum().item()
-
-        epoch_loss = running_loss / total
-        epoch_accuracy = correct / total
-
-        train_losses.append(epoch_loss)
-        train_accuracies.append(epoch_accuracy)
-        val_loss, val_accuracy = evaluate(model, val_loader, criterion)
-        val_losses.append(val_loss)
-        val_accuracies.append(val_accuracy)
-
-        print(
-            f"Epoch {epoch + 1}/{num_epochs}, Train Loss: {epoch_loss:.4f}, Train Accuracy: {epoch_accuracy:.4f}, Validation Loss: {val_loss:.4f}, Validation Accuracy: {val_accuracy:.4f}")
-
-    return train_losses, val_losses, train_accuracies, val_accuracies
-
-
-# Evaluation function
-def evaluate(model, val_loader, criterion):
-    model.eval()
-    val_loss = 0.0
-    correct = 0
-    total = 0
-
-    with torch.no_grad():
-        for inputs, labels in val_loader:
-            inputs, labels = inputs.to(device), labels.to(device)
-            outputs = model(inputs)
-            loss = criterion(outputs, labels)
-
-            val_loss += loss.item() * inputs.size(0)
-            _, predicted = torch.max(outputs, 1)
-            total += labels.size(0)
-            correct += (predicted == labels).sum().item()
-
-    val_loss /= total
-    val_accuracy = correct / total
-    return val_loss, val_accuracy
-
-
-if __name__ == "__main__":
-    foldername = 'malware-images-large'
-
-    num_classes = len(DATASET_CLASSES[dataset_type])
-    batch_size = 32
-    num_epochs = 10
-    learning_rate = 0.001
-
-    # Load data
-    train_loader, val_loader, test_loader = load_data(foldername)
-
-    # Initialize the model
-    weight_path = 'C:/Users/zheng/OneDrive/Desktop/CurrentSem/COMP6721(AppliedAI/Project/Malware-Detection-CNN/trained models/malware_moderate_efficientnetb0.pth'
-    efficientnet_model = EfficientNetModel(num_classes=num_classes, pretrained=False, weight_path=weight_path)
-    model = efficientnet_model.get_model().to(device)
-
-    path = 'C:/data/malware_tran_large_efficientnetb0.pth'
-
-    # Define loss function and optimizer
-    criterion = nn.CrossEntropyLoss()
-    optimizer = optim.Adam(model.parameters(), lr=learning_rate)
-
-    train_losses, val_losses, train_accuracies, val_accuracies = train(model, train_loader, val_loader, criterion,
-                                                                       optimizer, num_epochs)
-
-    save_model(model, path)
-
-    trained_model = load_model(path, len(DATASET_CLASSES[dataset_type]))
-
-    plot_metrics(train_losses, val_losses, train_accuracies, val_accuracies)
-
-    # Call the evaluation function
-    test_loss, test_accuracy = evaluate_on_test_data(model, test_loader, criterion)
-
-    # Print the results
-    print(f"Test Loss: {test_loss:.4f}, Test Accuracy: {test_accuracy:.2f}%")
-
-    y_pred, y_true = get_predictions(model, test_loader)
-    unique_true = np.unique(y_true)
-    unique_pred = np.unique(y_pred)
-
-    print("Unique classes in y_true:", unique_true)
-    print("Unique classes in y_pred:", unique_pred)
-
-    # Plot confusion matrix
-    plot_confusion_matrix(y_true, y_pred, DATASET_CLASSES[dataset_type])
-
-
-
