--- tran_Resnet_moderate_to_large.py
@@ -1,246 +1,0 @@
-import sys
-import os
-import torch
-import torch.nn as nn
-import torch.optim as optim
-from torch.utils.data import DataLoader, TensorDataset
-import numpy as np
-from torchvision.models import resnet18, ResNet18_Weights
-sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "../..")))
-from src.preprocessing.constants import *
-from src.evaluation.evaluate import load_model,plot_metrics,get_predictions,plot_confusion_matrix
-
-
-device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
-
-dataset_type = DatasetType.LARGE #.LARGE, .MODERATE, .SMALL
-
-class ResNet18(nn.Module):
-    def __init__(self, num_classes):
-        super(ResNet18, self).__init__()
-        self.model = resnet18(weights=None)
-        self.model.fc =  nn.Sequential(nn.Linear(self.model.fc.in_features, num_classes))
-
-    def forward(self, x):
-        return self.model(x)
-
-def load_data(foldername):
-    X_train = np.load('C:/Users/zheng/OneDrive/Desktop/CurrentSem/COMP6721(AppliedAI/Project/Malware-Detection-CNN/src/preprocessing/data/processed/malware-images-large/X_train.npy')
-    y_train = np.load('C:/Users/zheng/OneDrive/Desktop/CurrentSem/COMP6721(AppliedAI/Project/Malware-Detection-CNN/src/preprocessing/data/processed/malware-images-large/y_train.npy')
-    X_val = np.load('C:/Users/zheng/OneDrive/Desktop/CurrentSem/COMP6721(AppliedAI/Project/Malware-Detection-CNN/src/preprocessing/data/processed/malware-images-large/X_val.npy')
-    y_val = np.load('C:/Users/zheng/OneDrive/Desktop/CurrentSem/COMP6721(AppliedAI/Project/Malware-Detection-CNN/src/preprocessing/data/processed/malware-images-large/y_val.npy')
-
-    # Stack the grayscale image data into 3 channels to match ResNet-18’s expected input format
-    X_train_tensor = torch.tensor(X_train, dtype=torch.float32).unsqueeze(1).repeat(1, 3, 1,1)  # (num_images, 3, 224, 224)
-    y_train_tensor = torch.tensor(y_train, dtype=torch.long)
-
-    X_val_tensor = torch.tensor(X_val, dtype=torch.float32).unsqueeze(1).repeat(1, 3, 1,1)  # (num_images_val, 3, 224, 224)
-    y_val_tensor = torch.tensor(y_val, dtype=torch.long)
-
-    train_dataset = TensorDataset(X_train_tensor, y_train_tensor)
-    val_dataset = TensorDataset(X_val_tensor, y_val_tensor)
-
-    train_loader = DataLoader(train_dataset, batch_size=32, shuffle=True)
-    val_loader = DataLoader(val_dataset, batch_size=32, shuffle=False)
-
-    X_test = np.load('C:/Users/zheng/OneDrive/Desktop/CurrentSem/COMP6721(AppliedAI/Project/Malware-Detection-CNN/src/preprocessing/data/processed/malware-images-large/X_test.npy')
-    y_test = np.load('C:/Users/zheng/OneDrive/Desktop/CurrentSem/COMP6721(AppliedAI/Project/Malware-Detection-CNN/src/preprocessing/data/processed/malware-images-large/y_test.npy')  # Adjust path to your .npy file
-
-    # Convert numpy arrays to PyTorch tensors
-    X_test_tensor = torch.tensor(X_test, dtype=torch.float32).unsqueeze(1).repeat(1, 3, 1, 1)
-    y_test_tensor = torch.tensor(y_test, dtype=torch.long)
-
-    # Move the test data to the same device
-    X_test_tensor = X_test_tensor.to(device)
-    y_test_tensor = y_test_tensor.to(device)
-
-    test_data = TensorDataset(X_test_tensor, y_test_tensor)
-    test_loader = DataLoader(test_data, batch_size=32, shuffle=False)
-
-
-    # Check the count of each class based on y_test labels
-    unique_classes, class_counts = np.unique(y_test, return_counts=True)
-
-    print("Class distribution in X_test based on y_test labels:")
-    for cls, count in zip(unique_classes, class_counts):
-        print(f"Class {cls}: {count} samples in X_test")
-
-    unique_classes_train, class_counts_train = np.unique(y_train, return_counts=True)
-
-    print("Class distribution in X_train based on y_train labels:")
-    for cls, count in zip(unique_classes_train, class_counts_train):
-        print(f"Class {cls}: {count} samples in X_train")
-    return train_loader, val_loader, test_loader
-
-
-def train_model(model, train_loader, optimizer, criterion):
-    model.train()
-    running_loss = 0.0
-    correct_predictions = 0
-    total_predictions = 0
-
-    for inputs, labels in train_loader:
-        if torch.cuda.is_available():
-            inputs, labels = inputs.cuda(), labels.cuda()
-            model.cuda()
-
-        optimizer.zero_grad()
-        outputs = model(inputs)
-        loss = criterion(outputs, labels)
-        loss.backward()
-        optimizer.step()
-
-        running_loss += loss.item()
-        _, predicted = torch.max(outputs, 1)
-        correct_predictions += (predicted == labels).sum().item()
-        total_predictions += labels.size(0)
-
-    epoch_loss = running_loss / len(train_loader)
-    epoch_accuracy = correct_predictions / total_predictions
-    return epoch_loss, epoch_accuracy
-
-
-def validate_model(model, val_loader, criterion):
-    model.eval()
-    val_loss = 0.0
-    correct_predictions = 0
-    total_predictions = 0
-
-    with torch.no_grad():
-        for inputs, labels in val_loader:
-            if torch.cuda.is_available():
-                inputs, labels = inputs.cuda(), labels.cuda()
-
-            outputs = model(inputs)
-            loss = criterion(outputs, labels)
-            val_loss += loss.item()
-
-            _, predicted = torch.max(outputs, 1)
-            correct_predictions += (predicted == labels).sum().item()
-            total_predictions += labels.size(0)
-
-    val_loss /= len(val_loader)
-    val_accuracy = correct_predictions / total_predictions
-    return val_loss, val_accuracy
-
-
-def train_loop(model,train_loader,val_loader,criterion,num_epochs=10):
-
-    optimizer = optim.Adam(model.parameters(), lr=1e-4)
-
-    train_losses, val_losses = [], []
-    train_accuracies, val_accuracies = [], []
-    y_train_true, y_train_pred = [], []
-    y_val_true, y_val_pred = [], []
-
-    for epoch in range(num_epochs):
-        train_loss, train_accuracy = train_model(model, train_loader, optimizer, criterion)
-        val_loss, val_accuracy = validate_model(model, val_loader, criterion)
-
-        train_losses.append(train_loss)
-        val_losses.append(val_loss)
-        train_accuracies.append(train_accuracy)
-        val_accuracies.append(val_accuracy)
-
-        print(f"Epoch {epoch + 1}/{num_epochs}, "
-              f"Train Loss: {train_loss:.4f}, Train Accuracy: {train_accuracy:.4f}, "
-              f"Val Loss: {val_loss:.4f}, Val Accuracy: {val_accuracy:.4f}")
-
-    return train_losses, val_losses, train_accuracies, val_accuracies
-
-def evaluate_on_test_data(model, test_loader, criterion):
-
-    # Move the model to the selected device
-    model = model.to(device)
-
-    # Set the model to evaluation mode
-    model.eval()
-
-    # Initialize variables to track loss and accuracy
-    test_loss = 0.0
-    correct = 0
-    total = 0
-
-    # Disable gradient calculation
-    with torch.no_grad():
-        for inputs, labels in test_loader:
-            # Forward pass
-            outputs = model(inputs)
-
-            # Calculate the loss
-            loss = criterion(outputs, labels)
-            test_loss += loss.item() * inputs.size(0)  # Accumulate loss
-
-            # Calculate the number of correct predictions
-            _, predicted = torch.max(outputs, 1)
-            total += labels.size(0)
-            correct += (predicted == labels).sum().item()
-
-    # Calculate average loss and accuracy
-    avg_loss = test_loss / total
-    accuracy = correct / total * 100
-
-    return avg_loss, accuracy
-
-def save_model(model, path="model.pth"):
-    torch.save(model.state_dict(), path)
-    print(f"Model saved to {path}")
-
-
-# Run the main function
-if __name__ == "__main__":
-    # print(DATASET_CLASSES[dataset_type])
-    foldername = 'malware-images-large'
-    model = ResNet18(num_classes=len(DATASET_CLASSES[dataset_type]))
-
-    # Example criterion for classification (Cross-Entropy Loss)
-    criterion = torch.nn.CrossEntropyLoss()
-
-    train_loader, val_loader,test_loader = load_data(foldername)
-
-    # Load the saved weights
-    checkpoint_path = 'C:/Users/zheng/OneDrive/Desktop/CurrentSem/COMP6721(AppliedAI/Project/Malware-Detection-CNN/trained models/malware_moderate_resnet18.pth'
-    checkpoint = torch.load(checkpoint_path, map_location=torch.device('cpu'))
-
-    # Remove the mismatched fc layer weights from the checkpoint
-    fc_layer_keys = [key for key in checkpoint.keys() if key.startswith("model.fc")]
-    for key in fc_layer_keys:
-        print(f"Removing mismatched key: {key}")
-        del checkpoint[key]
-
-    # Load weight to model
-    load_result  = model.load_state_dict(checkpoint, strict=False)
-
-    missing = load_result.missing_keys
-    unexpected = load_result.unexpected_keys
-
-    print("Missing layers:", missing)
-    print("Unexpected layers:", unexpected)
-    print("Weights loaded successfully, excluding incompatible layers.")
-
-    train_losses, val_losses, train_accuracies, val_accuracies=train_loop(model,train_loader,val_loader,criterion)
-
-    path = 'C:/data/malware_tran_large_resnet18.pth'
-
-
-    save_model(model, path)
-
-    trained_model = load_model(path,len(DATASET_CLASSES[dataset_type]))
-
-    plot_metrics(train_losses, val_losses, train_accuracies, val_accuracies)
-
-    # Call the evaluation function
-    test_loss, test_accuracy = evaluate_on_test_data(model,test_loader, criterion)
-
-    # Print the results
-    print(f"Test Loss: {test_loss:.4f}, Test Accuracy: {test_accuracy:.2f}%")
-
-    y_pred, y_true = get_predictions(model, test_loader)
-    unique_true = np.unique(y_true)
-    unique_pred = np.unique(y_pred)
-
-    print("Unique classes in y_true:", unique_true)
-    print("Unique classes in y_pred:", unique_pred)
-
-    # Plot confusion matrix
-    plot_confusion_matrix(y_true, y_pred, DATASET_CLASSES[dataset_type])
